+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # v1.3.0
 
 IMPROVEMENTS
+- Add `UPLOAD_BY_PART` (default: true) option for improve upload/download concurrency fix [#324](https://github.com/AlexAkulov/clickhouse-backup/issues/324)
 - Add ARM support for Docker images and pre-compiled binary files, fix [#312](https://github.com/AlexAkulov/clickhouse-backup/issues/312)
 - KeepRemoteBackups should respect differential backups, fix [#111](https://github.com/AlexAkulov/clickhouse-backup/issues/111)
 - Add SFTP_DEBUG, fix [#335](https://github.com/AlexAkulov/clickhouse-backup/issues/335)